Development and production mode in electron main process.
env.ts is created just before transpile main.ts to main.js. Quit that way and decide to use dotenv to know if development mode is turned on. Prepare .env for development and .env.prod for production in the root directory. And when building the app, .env.prod will be copied into dist directory as .env.